all: clean up lint identified by ruleguard
Ruleguard run with ruleguard -c=0 -rules ruleguard.rules.go ./... from
https://github.com/dgryski/semgrep-go with the following rules inactivated
to reduce noise:

- unconvert for floating point values
- oddcomparisons
- floateq

+++ graph/encoding/dot/encode_test.go
@@ -388,7 +388,7 @@ func undirectedSubGraphFrom(g []intset, s map[int64][]intset) graph.Graph {
 				sub.SetEdge(ce)
 			}
 		}
-		subs[i] = subGraph{id: int64(i), Graph: sub}
+		subs[i] = subGraph{id: i, Graph: sub}
 		base += int64(len(sg))
 	}
